Disable inflection point search for small datasets in flexiplex-filter
There are often times when it is desirable to have pipelines which work automatically on smaller datasets, especially
'example' or demo ones.

In these cases, inflection point searching is often not desirable due to the size of the datasets. In the cases where
it _is_ useful, it is strongly recommended that users first review the data manually and configure flexiplex-filter using
the -u and -l parameters. Hence, this commit disables the flexiplex-filter inflection point discovery automatically for
cases where there are under 50 unique counts, unless the -u and -l parameters are manually specified.